more DocPage refactors
